#37 refactoring: apartmentservice 리팩토링
